# src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Course API",
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/public",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/course": {
      "post": {
        "description": "create course",
        "summary": "create course",
        "tags": [
          "course"
        ],
        "operationId": "createcourse",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createcourserequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/%7B%7BBASE%7D%7D~1public~1course"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Fri, 03 Feb 2023 16:40:14 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    },
    "/course/filtered": {
      "post": {
        "description": "get all courses",
        "summary": "get all  courses",
        "tags": [
          "course"
        ],
        "operationId": "getallcourses",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/getallcoursesrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/%7B%7BBASE%7D%7D~1public~1course~1filtered"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "result": {
                  "offset": 0,
                  "limit": 3,
                  "totalRecords": 2,
                  "records": [
                    {
                      "id": "d97d94d3-83fb-427d-8326-c0474bc0eec0",
                      "title": "kompuktorni clases",
                      "description": "learn kompuktory",
                      "generalInformation": [
                        "all information about kompuktory"
                      ],
                      "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                      "updatedBy": null,
                      "createdAt": "2023-02-03T16:40:14.000Z",
                      "updatedAt": "2023-02-03T16:40:14.000Z"
                    },
                    {
                      "id": "9b9b26dc-8f3b-4b8b-a269-0586b07ad43d",
                      "title": "kompuktorni clases for dumps",
                      "description": "learn kompuktory",
                      "generalInformation": [
                        "all information about kompuktory"
                      ],
                      "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                      "updatedBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                      "createdAt": "2023-02-03T15:26:14.000Z",
                      "updatedAt": "2023-02-03T15:43:04.000Z"
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "745"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2e9-xiZOY1Kun+w6hXRJ1HHRzdv4rUU\""
              },
              "Date": {
                "type": "string",
                "default": "Fri, 03 Feb 2023 17:03:56 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/lesson/{lessonId}/course/{courseId}": {
      "get": {
        "summary": "get lesson by Id",
        "tags": [
          "lesson"
        ],
        "operationId": "getlessonbyId",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getlessonbyId"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "result": {
                  "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
                  "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                  "title": "kompuktory",
                  "lessonNumber": 1,
                  "description": "learn firs kompuktor",
                  "information": [
                    "information about kompuktory"
                  ],
                  "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                  "updatedBy": null,
                  "createdAt": "2023-02-04T17:50:09.000Z",
                  "updatedAt": "2023-02-04T17:50:09.000Z"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "389"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"185-jxq++j7e5fzHF3k4OwbAg52XFVc\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:49:32 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "update lesson",
        "tags": [
          "lesson"
        ],
        "operationId": "updatelesson",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updatelessonrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/updatelesson"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:54:21 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "summary": "delete lesson",
        "tags": [
          "lesson"
        ],
        "operationId": "deletelesson",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/deletelesson"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:57:23 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
...
    "createcourserequest": {
      "title": "createcourserequest",
      "example": {
        "title": "kompuktorni clases",
        "description": "learn kompuktory",
        "instructorIds": [
          "0250ad05-8c3c-49c0-9fc9-c50caf6b415e"
        ],
        "generalInformation": [
          "all information about kompuktory"
        ],
        "lessons": [
          {
            "title": "kompuktory",
            "description": "learn firs kompuktor",
            "information": [
              "information about kompuktory"
            ]
          },
          {
            "title": "kompuktory 2",
            "description": "learn second kompuktor",
            "information": [
              "information about kompuktory"
            ]
          },
          {
            "title": "kompuktory 3",
            "description": "learn kompuktor of 3 gen",
            "information": [
              "information about kompuktory"
            ]
          },
          {
            "title": "kompuktory 4",
            "description": "learn kompuktor of 4 gen",
            "information": [
              "information about kompuktory"
            ]
          },
          {
            "title": "kompuktory 5",
            "description": "learn kompuktor of 5 gen",
            "information": [
              "information about kompuktory"
            ]
          }
        ]
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "instructorIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "generalInformation": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "lessons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Lesson"
          }
        }
      },
      "required": [
        "title",
        "description",
        "instructorIds",
        "generalInformation",
        "lessons"
      ]
    },
    "Lesson": {
      "title": "Lesson",
      "example": {
        "title": "kompuktory",
        "description": "learn firs kompuktor",
        "information": [
          "information about kompuktory"
        ]
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "title",
        "description",
        "information"
      ]
    },
...
    "getallcoursesrequest": {
      "title": "getallcoursesrequest",
      "example": {
        "offset": 0,
        "limit": 3,
        "orderBy": "title",
        "orderDirection": "asc"
      },
      "type": "object",
      "properties": {
        "offset": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "orderBy": {
          "type": "string"
        },
        "orderDirection": {
          "type": "string"
        }
      },
      "required": [
        "offset",
        "limit",
        "orderBy",
        "orderDirection"
      ]
    },
...
    "Result11": {
      "title": "Result11",
      "example": {
        "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
        "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
        "title": "kompuktory",
        "lessonNumber": 1,
        "description": "learn firs kompuktor",
        "information": [
          "information about kompuktory"
        ],
        "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
        "updatedBy": null,
        "createdAt": "2023-02-04T17:50:09.000Z",
        "updatedAt": "2023-02-04T17:50:09.000Z"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "courseId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "lessonNumber": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdBy": {
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "courseId",
        "title",
        "lessonNumber",
        "description",
        "information",
        "createdBy",
        "updatedBy",
        "createdAt",
        "updatedAt"
      ]
    },
    "getlessonbyId": {
      "title": "getlessonbyId",
      "example": {
        "success": true,
        "result": {
          "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
          "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
          "title": "kompuktory",
          "lessonNumber": 1,
          "description": "learn firs kompuktor",
          "information": [
            "information about kompuktory"
          ],
          "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
          "updatedBy": null,
          "createdAt": "2023-02-04T17:50:09.000Z",
          "updatedAt": "2023-02-04T17:50:09.000Z"
        }
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "result": {
          "$ref": "#/definitions/Result11"
        }
      },
      "required": [
        "success",
        "result"
      ]
    },
    "updatelessonrequest": {
      "title": "updatelessonrequest",
      "example": {
        "title": "kompuktory",
        "description": "!!!!!!!!!!!!!!!",
        "information": [
          "information about kompuktory"
        ]
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "title",
        "description",
        "information"
      ]
    },
    "updatelesson": {
      "title": "updatelesson",
      "example": {
        "success": true
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      },
      "required": [
        "success"
      ]
    },
    "deletelesson": {
      "title": "deletelesson",
      "example": {
        "success": true
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      },
      "required": [
        "success"
      ]
    }
  },
...
}